test_project/requirements/base.txt
hiredis
django_redis
psycopg2-binary==2.8.4
git+git://github.com/rajeshyogeshwar/django-tenant-schemas-q